Atualizado

+++ implementacao/controllers/DeclaracaoController.php
@@ -3,16 +3,17 @@
 namespace app\controllers;
 
 use Yii;
-use app\models\Area;
-use app\models\AreaSearch;
+use app\models\Declaracao;
+use app\models\DeclaracaoSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use \yii\web\ForbiddenHttpException;
 
 /**
- * AreaController implements the CRUD actions for Area model.
+ * DeclaracaoController implements the CRUD actions for Declaracao model.
  */
-class AreaController extends Controller
+class DeclaracaoController extends Controller
 {
     /**
      * {@inheritdoc}
@@ -30,53 +31,66 @@ class AreaController extends Controller
     }
 
     /**
-     * Lists all Area models.
+     * Lists all Declaracao models.
      * @return mixed
      */
     public function actionIndex()
     {
-        $searchModel = new AreaSearch();
+        if (!Yii::$app->user->isGuest && Yii::$app->user->identity->role == 2) { 
+        $searchModel = new DeclaracaoSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        } else {
+            throw new \yii\web\ForbiddenHttpException('Você não tem permissão para acessar esta página.');
+        }
     }
 
     /**
-     * Displays a single Area model.
+     * Displays a single Declaracao model.
      * @param integer $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->can('coordenador')) {
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        
+        } else {
+            throw new ForbiddenHttpException('Você não tem permissão para acessar esta página.');
+        }
     }
 
     /**
-     * Creates a new Area model.
+     * Creates a new Declaracao model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
     public function actionCreate()
     {
-        $model = new Area();
-
+        if (Yii::$app->user->can('coordenador')) {
+        $model = new Declaracao();
+        $model->dataEmissao = new \yii\db\Expression('NOW()');
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            return $this->redirect(['index']);
         }
 
         return $this->render('create', [
             'model' => $model,
         ]);
+        } else {
+            throw new ForbiddenHttpException('Você não tem permissão para acessar esta página.');
+        }
     }
 
     /**
-     * Updates an existing Area model.
+     * Updates an existing Declaracao model.
      * If update is successful, the browser will be redirected to the 'view' page.
      * @param integer $id
      * @return mixed
@@ -84,6 +98,7 @@ class AreaController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->can('coordenador')) {
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -93,10 +108,13 @@ class AreaController extends Controller
         return $this->render('update', [
             'model' => $model,
         ]);
+        } else {
+            throw new ForbiddenHttpException('Você não tem permissão para acessar esta página.');
+        }
     }
 
     /**
-     * Deletes an existing Area model.
+     * Deletes an existing Declaracao model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param integer $id
      * @return mixed
@@ -104,21 +122,25 @@ class AreaController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->can('coordenador')) {
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException('Você não tem permissão para acessar esta página.');
+        }
     }
 
     /**
-     * Finds the Area model based on its primary key value.
+     * Finds the Declaracao model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
      * @param integer $id
-     * @return Area the loaded model
+     * @return Declaracao the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
     protected function findModel($id)
     {
-        if (($model = Area::findOne($id)) !== null) {
+        if (($model = Declaracao::findOne($id)) !== null) {
             return $model;
         }
